# apps/api/openapi/releases/1.0.0.json
{
  "components": {
    "schemas": {
      "AcceptedResponseDto": {
        "additionalProperties": false,
        "properties": { "accepted": { "example": true, "type": "boolean" } },
        "required": ["accepted"],
        "type": "object"
      },
      "ApiErrorDetailDto": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "enum": [
              "VALIDATION_FAILED",
              "UNAUTHENTICATED",
              "FORBIDDEN",
              "NOT_FOUND",
              "CONFLICT",
              "VERSION_MISMATCH",
              "IDEMPOTENCY_KEY_REUSED",
              "RATE_LIMITED",
              "SERVICE_UNAVAILABLE",
              "INTERNAL_ERROR"
            ],
            "example": "VALIDATION_FAILED",
            "type": "string"
          },
          "message": {
            "example": "The request could not be processed.",
            "type": "string"
          },
          "requestId": {
            "example": "request_123",
            "maxLength": 128,
            "type": "string"
          }
        },
        "required": ["code", "message", "requestId"],
        "type": "object"
      },
      "ApiErrorResponseDto": {
        "additionalProperties": false,
        "properties": {
          "error": { "$ref": "#/components/schemas/ApiErrorDetailDto" }
        },
        "required": ["error"],
        "type": "object"
      },
      "AuthenticatedUserResponseDto": {
        "additionalProperties": false,
        "properties": {
          "user": { "$ref": "#/components/schemas/UserResponseDto" }
        },
        "required": ["user"],
        "type": "object"
      },
      "ChangePasswordRequestDto": {
        "additionalProperties": false,
        "properties": {
          "currentPassword": {
            "format": "password",
            "maxLength": 128,
            "type": "string",
            "writeOnly": true
          },
          "newPassword": {
            "format": "password",
            "maxLength": 128,
            "minLength": 12,
            "type": "string",
            "writeOnly": true
          }
        },
        "required": ["currentPassword", "newPassword"],
        "type": "object"
      },
      "ConfirmRecentAuthRequestDto": {
        "additionalProperties": false,
        "properties": {
          "password": {
            "format": "password",
            "maxLength": 128,
            "type": "string",
            "writeOnly": true
          }
        },
        "required": ["password"],
        "type": "object"
      },
      "EmailRequestDto": {
        "additionalProperties": false,
        "properties": {
          "email": { "format": "email", "maxLength": 254, "type": "string" }
        },
        "required": ["email"],
        "type": "object"
      },
...
      "LoginRequestDto": {
        "additionalProperties": false,
        "properties": {
          "email": { "format": "email", "maxLength": 254, "type": "string" },
          "password": {
            "format": "password",
            "maxLength": 128,
            "minLength": 12,
            "type": "string",
            "writeOnly": true
          }
        },
        "required": ["email", "password"],
        "type": "object"
      },
      "RefreshResponseDto": {
        "additionalProperties": false,
        "properties": { "refreshed": { "example": true, "type": "boolean" } },
        "required": ["refreshed"],
        "type": "object"
      },
      "RegisterRequestDto": {
        "additionalProperties": false,
        "properties": {
          "email": { "format": "email", "maxLength": 254, "type": "string" },
          "password": {
            "format": "password",
            "maxLength": 128,
            "minLength": 12,
            "type": "string",
            "writeOnly": true
          }
        },
        "required": ["email", "password"],
        "type": "object"
      },
      "ResetPasswordRequestDto": {
        "additionalProperties": false,
        "properties": {
          "newPassword": {
            "format": "password",
            "maxLength": 128,
            "minLength": 12,
            "type": "string",
            "writeOnly": true
          },
          "token": {
            "maxLength": 1024,
            "minLength": 32,
            "type": "string",
            "writeOnly": true
          }
        },
        "required": ["token", "newPassword"],
        "type": "object"
      },
      "SessionListResponseDto": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": { "$ref": "#/components/schemas/SessionResponseDto" },
            "type": "array"
          }
        },
        "required": ["items"],
        "type": "object"
      },
...
      "UserResponseDto": {
        "additionalProperties": false,
        "properties": {
          "createdAt": { "format": "date-time", "type": "string" },
          "email": { "format": "email", "maxLength": 254, "type": "string" },
          "emailVerifiedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "id": { "format": "uuid", "type": "string" },
          "role": { "enum": ["user", "admin"], "type": "string" },
          "status": {
            "enum": ["pending_verification", "active", "suspended", "deleted"],
            "type": "string"
          },
          "updatedAt": { "format": "date-time", "type": "string" },
          "version": { "minimum": 0, "type": "number" }
        },
        "required": [
          "id",
          "email",
          "status",
          "role",
          "version",
          "emailVerifiedAt",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
      "VerifyEmailRequestDto": {
        "additionalProperties": false,
        "properties": {
          "token": {
            "maxLength": 1024,
            "minLength": 32,
            "type": "string",
            "writeOnly": true
          }
        },
        "required": ["token"],
        "type": "object"
      }
    },
    "securitySchemes": {
      "csrfCookie": {
        "in": "cookie",
        "name": "__Host-cs_csrf",
        "type": "apiKey"
      },
      "csrfHeader": {
        "description": "Session-bound CSRF token for unsafe cookie-auth requests.",
        "in": "header",
        "name": "X-CSRF-Token",
        "type": "apiKey"
      },
      "refreshCookie": {
        "in": "cookie",
        "name": "__Host-cs_refresh",
        "type": "apiKey"
      },
      "sessionCookie": {
        "in": "cookie",
        "name": "__Host-cs_access",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Versioned Cornerstone identity and platform API contract.",
    "title": "Cornerstone API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login",
        "parameters": [
          {
            "description": "Session-bound CSRF token.",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequestDto" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticatedUserResponseDto"
                }
              }
            },
            "description": "",
            "headers": {
              "Set-Cookie": { "description": "Host-only auth cookies." }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          }
        },
        "security": [{ "csrfCookie": [], "csrfHeader": [] }],
        "summary": "",
        "tags": ["Authentication"],
        "x-cornerstone-authorization": {
          "owner": "backend",
          "ownership": "none",
          "permission": null,
          "roles": []
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "logout",
        "parameters": [
          {
            "description": "Session-bound CSRF token.",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "204": {
            "description": "",
            "headers": {
              "Set-Cookie": { "description": "Expired auth cookies." }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          }
        },
        "security": [
          { "csrfCookie": [], "csrfHeader": [], "sessionCookie": [] }
        ],
        "summary": "",
        "tags": ["Authentication"],
        "x-cornerstone-authorization": {
          "owner": "backend",
          "ownership": "self",
          "permission": "session:revoke",
          "roles": ["user", "admin"]
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "getCurrentUser",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticatedUserResponseDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          }
        },
        "security": [{ "sessionCookie": [] }],
        "summary": "",
        "tags": ["Authentication"],
        "x-cornerstone-authorization": {
          "owner": "backend",
          "ownership": "self",
          "permission": "profile:read",
          "roles": ["user", "admin"]
        }
      }
    },
    "/api/v1/auth/password/change": {
      "post": {
        "operationId": "changePassword",
        "parameters": [
          {
            "description": "Session-bound CSRF token.",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": { "description": "" },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          }
        },
        "security": [
          { "csrfCookie": [], "csrfHeader": [], "sessionCookie": [] }
        ],
        "summary": "",
        "tags": ["Authentication"],
        "x-cornerstone-authorization": {
          "owner": "backend",
          "ownership": "self",
          "permission": "profile:update",
          "roles": ["user", "admin"]
        }
      }
    },
    "/api/v1/auth/password/forgot": {
      "post": {
        "operationId": "requestPasswordReset",
        "parameters": [
          {
            "description": "Session-bound CSRF token.",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EmailRequestDto" }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AcceptedResponseDto" }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          }
        },
        "security": [{ "csrfCookie": [], "csrfHeader": [] }],
        "summary": "",
        "tags": ["Authentication"],
        "x-cornerstone-authorization": {
          "owner": "backend",
          "ownership": "none",
          "permission": null,
          "roles": []
        }
      }
    },
    "/api/v1/auth/password/reset": {
      "post": {
        "operationId": "resetPassword",
        "parameters": [
          {
            "description": "Session-bound CSRF token.",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": { "description": "" },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          }
        },
        "security": [{ "csrfCookie": [], "csrfHeader": [] }],
        "summary": "",
        "tags": ["Authentication"],
        "x-cornerstone-authorization": {
          "owner": "backend",
          "ownership": "none",
          "permission": null,
          "roles": []
        }
      }
    },
    "/api/v1/auth/recent-auth": {
      "post": {
        "operationId": "confirmRecentAuthentication",
        "parameters": [
          {
            "description": "Session-bound CSRF token.",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmRecentAuthRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": { "description": "" },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          }
        },
        "security": [
          { "csrfCookie": [], "csrfHeader": [], "sessionCookie": [] }
        ],
        "summary": "",
        "tags": ["Authentication"],
        "x-cornerstone-authorization": {
          "owner": "backend",
          "ownership": "self",
          "permission": "profile:update",
          "roles": ["user", "admin"]
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "refreshSession",
        "parameters": [
          {
            "description": "Session-bound CSRF token.",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RefreshResponseDto" }
              }
            },
            "description": "",
            "headers": {
              "Set-Cookie": { "description": "Rotated host-only auth cookies." }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          }
        },
        "security": [
          { "csrfCookie": [], "csrfHeader": [], "refreshCookie": [] }
        ],
        "summary": "",
        "tags": ["Authentication"],
        "x-cornerstone-authorization": {
          "owner": "backend",
          "ownership": "self",
          "permission": null,
          "roles": []
        }
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "operationId": "register",
        "parameters": [
          {
            "description": "Session-bound CSRF token.",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequestDto" }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AcceptedResponseDto" }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          }
        },
        "security": [{ "csrfCookie": [], "csrfHeader": [] }],
        "summary": "",
        "tags": ["Authentication"],
        "x-cornerstone-authorization": {
          "owner": "backend",
          "ownership": "none",
          "permission": null,
          "roles": []
        }
      }
    },
    "/api/v1/auth/sessions": {
      "delete": {
        "operationId": "revokeAllSessions",
        "parameters": [
          {
            "description": "Session-bound CSRF token.",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "204": { "description": "" },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          }
        },
        "security": [
          { "csrfCookie": [], "csrfHeader": [], "sessionCookie": [] }
        ],
        "summary": "",
        "tags": ["Authentication"],
        "x-cornerstone-authorization": {
          "owner": "backend",
          "ownership": "self",
          "permission": "session:revoke",
          "roles": ["user", "admin"]
        }
      },
      "get": {
        "operationId": "listSessions",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionListResponseDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          }
        },
        "security": [{ "sessionCookie": [] }],
        "summary": "",
        "tags": ["Authentication"],
        "x-cornerstone-authorization": {
          "owner": "backend",
          "ownership": "self",
          "permission": "session:list",
          "roles": ["user", "admin"]
        }
      }
    },
    "/api/v1/auth/sessions/{sessionId}": {
      "delete": {
        "operationId": "revokeSession",
        "parameters": [
          {
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": { "format": "uuid", "type": "string" }
          },
          {
            "description": "Session-bound CSRF token.",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "204": { "description": "" },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          }
        },
        "security": [
          { "csrfCookie": [], "csrfHeader": [], "sessionCookie": [] }
        ],
        "summary": "",
        "tags": ["Authentication"],
        "x-cornerstone-authorization": {
          "owner": "backend",
          "ownership": "self",
          "permission": "session:revoke",
          "roles": ["user", "admin"]
        }
      }
    },
    "/api/v1/auth/verification/resend": {
      "post": {
        "operationId": "resendVerification",
        "parameters": [
          {
            "description": "Session-bound CSRF token.",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EmailRequestDto" }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AcceptedResponseDto" }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          }
        },
        "security": [{ "csrfCookie": [], "csrfHeader": [] }],
        "summary": "",
        "tags": ["Authentication"],
        "x-cornerstone-authorization": {
          "owner": "backend",
          "ownership": "none",
          "permission": null,
          "roles": []
        }
      }
    },
    "/api/v1/auth/verify-email": {
      "post": {
        "operationId": "verifyEmail",
        "parameters": [
          {
            "description": "Session-bound CSRF token.",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/VerifyEmailRequestDto" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AcceptedResponseDto" }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponseDto" }
              }
            },
            "description": ""
          }
        },
        "security": [{ "csrfCookie": [], "csrfHeader": [] }],
        "summary": "",
        "tags": ["Authentication"],
        "x-cornerstone-authorization": {
          "owner": "backend",
          "ownership": "none",
          "permission": null,
          "roles": []
        }
      }
    },
...
  },
  "security": [{ "sessionCookie": [] }],
  "servers": [{ "description": "API origin", "url": "/" }],
...
}